fix: refactoring categories and articles admin panel
