+++ components/footer.tsx
@@ -202,7 +202,7 @@ export default function FooterAnimated() {
                     </motion.div>
                 </div>
                 <div className="border-t border-gray-800 mt-8 pt-6 text-center text-gray-500 text-sm">
-                    <p>© {new Date().getFullYear()} Game Tested Tech. All rights reserved.</p>
+                    <p>© {new Date().getFullYear()} Game Tested Tech LTD. All rights reserved.</p>
                 </div>
             </div>
         </motion.footer>
